Audit Batch 2: track-record integrity + edge validation
2.1 Strangle risk: max_loss now a +/-15% stress-move tail estimate (floored
    at old 3x credit) instead of flat 3x that understated tail 10-30x;
    buying-power gate uses capital consumed (max_loss) for credit trades
    instead of the inverted can-we-afford-income check.
2.2 Partial exits: book on REAL fill (await fill, cancel+skip on timeout)
    with fill-derived P&L; no more estimated P&L recorded pre-fill with the
    underlying price as the option 'fill'.
2.3 Learning size multiplier: re-validate at the actual executed size when
    the multiplier scales up; fall back to gate-approved size on rejection.
2.4 Range-model CV: purge gap now defaults to the label horizon (was 5 vs
    30-day overlapping labels on the XGB/LGBM path) — kills the leakage
    inflating the accuracy that gates iron-condor entries.
2.5 Thompson decay: float counts (int truncation zeroed 1-win arms daily,
    preventing any signal accumulation at low trade volume).

Audit ref: PLAN.md Batch 2. Co-Authored-By: Claude Opus 4.8 <noreply@anthropic.com>

+++ src/ait/strategies/thompson.py
@@ -31,8 +31,12 @@ class StrategyArm:
     """One arm of the multi-armed bandit."""
 
     name: str
-    wins: int = 0
-    losses: int = 0
+    # Floats, not ints: daily decay must shrink counts fractionally.
+    # int truncation (`int(1 * 0.995) = 0`) zeroed any 1-win arm every day,
+    # so the bandit could never accumulate signal at low trade volume
+    # (audit 2026-07-07 item 2.5). Beta(alpha, beta) accepts floats natively.
+    wins: float = 0.0
+    losses: float = 0.0
     total_pnl: float = 0.0
 
     @property
@@ -44,7 +48,7 @@ class StrategyArm:
         return 1.0 + self.losses
 
     @property
-    def observations(self) -> int:
+    def observations(self) -> float:
         return self.wins + self.losses
 
     @property
@@ -129,8 +133,10 @@ class ThompsonSampler:
         Call this periodically (e.g., daily) to gradually forget old outcomes.
         """
         for arm in self._arms.values():
-            arm.wins = max(0, int(arm.wins * self._decay_factor))
-            arm.losses = max(0, int(arm.losses * self._decay_factor))
+            # Float decay — see StrategyArm: int truncation destroyed
+            # low-count arms (audit item 2.5).
+            arm.wins = max(0.0, arm.wins * self._decay_factor)
+            arm.losses = max(0.0, arm.losses * self._decay_factor)
 
         log.info("thompson_decay_applied", arms=len(self._arms))
         self._save_state()
